move user logic to separate file.
